Adjustment to give user access to granted domain only
